fix(Core): harden index snapshot, MFT parsing, delta counting, and alias filtering
- Snapshot: validate file length against header-derived total length so a
  truncated snapshot fails at open with InvalidDataException instead of an
  access violation later
- MftParser: add missing bounds checks in ParseDataRunsInto so a malformed
  $MFT record cannot read past the record buffer
- DeltaOverlay: count resurrected records (re-upsert after remove) so counts
  no longer drift per add/remove/re-add cycle; path building extracted to
  DeltaPathBuilder to stay under the per-file line limit
- NetworkIndex: guard Count against ObjectDisposedException like sibling
  methods; consolidate watcher apply methods into one helper
- Scheduler: break the periodic loop on ObjectDisposedException like
  SchedulerQueueRunner
- PathSearchFuzzy: widen keep * RefinementHeadroomFactor to long to avoid
  int overflow
- GlobToRegex: treat three or more consecutive stars as ** instead of
  silently dropping them
- PathMode gates: filter live alias generation by DisabledAliasIds, matching
  the baked-alias path

Includes MSTest regression coverage for the snapshot, MFT, delta, and glob
changes.

+++ Core/IndexV2/Search/PathMode/PathSearchFuzzy.cs
@@ -29,7 +29,8 @@ internal static class PathSearchFuzzy
 
         // See NameSearch: bounded by the index, and widened so a large limit cannot overflow.
         var keep = (int)Math.Min((long)Math.Max(limit, 8) * 8, snapshot.Count + delta.Added.Count);
-        var scanKeep = Math.Min(keep * RefinementHeadroomFactor, RefinementScanCap);
+        // Widen before multiplying so a large keep cannot overflow int (see NameSearch's identical fix).
+        var scanKeep = (int)Math.Min((long)keep * RefinementHeadroomFactor, RefinementScanCap);
         var topN = new FzfTopN(scanKeep);
 
         var lastSep = pathQuery.LastIndexOf(Path.DirectorySeparatorChar);
